Removing ill-conceived `bezier._doctest`.
Instead solving the re-use problem by adding git root to the PYTHONPATH
to make `tests.utils` importable.
