+++ .coveragerc
@@ -1,7 +1,5 @@
 [run]
 branch = True
-omit =
-    */bezier/_doctest.py
 
 [report]
 fail_under = 100
